go via header file and TMB macros
