+++ src/Makevars
@@ -1,1 +1,1 @@
-PKG_CXXFLAGS = -D TMBAD_FRAMEWORK
+PKG_CXXFLAGS = -D TMBAD_FRAMEWORK -D EIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS
