+++ src/mmrm.cpp
@@ -1,3 +1,4 @@
+#include "init.h"
 #include "covariance.h"
 
 // Definition:
